refactor: 환율 메서드 리턴타입 Double에서 Float으로 수정

+++ Backend/triplet/src/main/java/com/ssafy/triplet/exchange/util/ExchangeUtil.java
@@ -17,17 +17,17 @@ public class ExchangeUtil {
     private final WebClientUtil webClientUtil;
 
     // 통화코드로 현재 환율을 가지고 오는 메소드
-    public Double getExchangeRateByCurrencyCode(String currencyCode) {
+    public Float getExchangeRateByCurrencyCode(String currencyCode) {
         LocalDate now = LocalDate.now();
         List<ExchangeRate> response = webClientUtil
                 .getExchangeRate(now.format(DateTimeFormatter.ofPattern("yyyyMMdd")));
 
         for (ExchangeRate er : response) {
             if (er.getCurrencyCode().trim().equals(currencyCode.trim().toUpperCase())) {
-                return Double.parseDouble(er.getExchangeRate());
+                return Float.parseFloat(er.getExchangeRate());
             }
         }
-        return -1.0; // 못찾았으면 -1.0을 반환
+        return -1F; // 못찾았으면 -1.0을 반환
     }
 
     // 두 위도, 경도를 통한 계산
